+++ .gitignore
@@ -5,3 +5,6 @@ test.txt
 /analysis
 /venv
 /node_modules
+/src/server/static
+*.py[code]
+pycache
