+++ src/client/index.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import ReactDom from 'react-dom';
+import 'semantic-ui-css/semantic.min.css';
+import App from './App';
+
+ReactDom.render(<App />, document.getElementById('app'));
