+++ src/server/routes/api.py
@@ -1,0 +1,8 @@
+from src.server.controllers.api import EssayGrader
+
+api = [
+    {
+        'res': EssayGrader,
+        'url': '/api/essay/mark'
+    }
+]
